Split into smaller files and discrete functions the Voter Guide Possibility page. More testing needed.
